+++ docs/source/release_notes.rst
@@ -2,8 +2,8 @@ Release Notes
 -------------
 **Future Releases**
     * Enhancements
-        * Added the STLDecomposer :pr:`3741`
-        * Brought the PolynomialDecomposer up to parity with STLDecomposer :pr:`3768`
+        * Added the STL Decomposer :pr:`3741`
+        * Integrated STLDecomposer into AutoMLSearch for time series regression problems :pr:`3781`
     * Fixes
     * Changes
         * Cap Featuretools at < 1.15.0 :pr:`3775`